refactor: 예약, 예약대기 응답 Dto 추가
