+++ src/main/java/roomescape/payment/dto/request/TossPaymentConfirmRequest.java
@@ -1,4 +1,4 @@
-package roomescape.payment.client.dto.request;
+package roomescape.payment.dto.request;
 
 public record TossPaymentConfirmRequest(String orderId, Long amount, String paymentKey) {
 }
