+++ src/main/java/roomescape/payment/dto/response/TossPaymentResponse.java
@@ -1,8 +1,5 @@
-package roomescape.payment.client.dto.response;
+package roomescape.payment.dto.response;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
-
-import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 
 public record TossPaymentResponse(
