+++ src/main/java/roomescape/reservation/dto/request/ReservationRequest.java
@@ -1,7 +1,7 @@
 package roomescape.reservation.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import roomescape.payment.client.dto.request.TossPaymentConfirmRequest;
+import roomescape.payment.dto.request.TossPaymentConfirmRequest;
 
 import java.time.LocalDate;
 
@@ -13,7 +13,7 @@ public record ReservationRequest(
         String orderId,
         Long amount) {
 
-        public TossPaymentConfirmRequest tossPaymentConfirmRequest() {
-            return new TossPaymentConfirmRequest(this.orderId, this.amount, this.paymentKey);
-        }
+    public TossPaymentConfirmRequest tossPaymentConfirmRequest() {
+        return new TossPaymentConfirmRequest(this.orderId, this.amount, this.paymentKey);
+    }
 }
