+++ src/main/java/roomescape/reservation/dto/response/MyReservationResponse.java
@@ -1,14 +1,15 @@
 package roomescape.reservation.dto.response;
 
 import roomescape.reservation.domain.Reservation;
-import roomescape.waiting.domain.Waiting;
 
 public record MyReservationResponse(
         Long id,
         String theme,
         String date,
         String time,
-        String status
+        String status,
+        String paymentKey,
+        long amount
 ) {
     public static MyReservationResponse from(Reservation reservation) {
         return new MyReservationResponse(
@@ -16,17 +17,9 @@ public record MyReservationResponse(
                 reservation.getThemeName(),
                 reservation.getDate().toString(),
                 reservation.getReservationTime().toString(),
-                "예약"
-        );
-    }
-
-    public static MyReservationResponse fromWaiting(Waiting waiting, long rank) {
-        return new MyReservationResponse(
-                waiting.getId(),
-                waiting.getTheme().getName(),
-                waiting.getDate().toString(),
-                waiting.getTime().getStartAt().toString(),
-                String.valueOf(rank)
+                "예약",
+                reservation.getPayment().getPaymentKey(),
+                reservation.getPayment().getAmount()
         );
     }
 }
